FX-43 refactor: [BE] 파일명, 변수명 refactoring
