+++ redbox/src/main/java/fx/redbox/repository/donorcardrequest/DonorCardRequestRepository.java
@@ -1,11 +1,11 @@
-package fx.redbox.repository.request;
+package fx.redbox.repository.donorcardrequest;
 
 import fx.redbox.entity.donorCards.DonorCardRequest;
 import fx.redbox.entity.donorCards.DonorCardRequestForm;
 
 import java.util.List;
 
-public interface RequestRepository {
+public interface DonorCardRequestRepository {
 
     DonorCardRequest createDonorCardRequest(DonorCardRequest request, DonorCardRequestForm requestForm);
